% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkAddAlpha.r
\name{addAlpha}
\alias{addAlpha}
\alias{pkAddAlpha}
\title{Make color transparant}
\usage{
addAlpha(colors = c("red", "green", "blue"), alpha = 0.5)

pkAddAlpha(colors = c("red", "green", "blue"), alpha = 0.5)
}
\arguments{
\item{alpha}{transparency value}

\item{x}{(vector of) color(s)}
}
\value{
transparant (vector of) color(s)
}
\description{
Make color transparant
}
