% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseString.r
\name{collapseString}
\alias{collapseString}
\title{Collapse string}
\usage{
collapseString(x, delimiter = ",", quotes = F)
}
\arguments{
\item{x}{(vector of strings}

\item{delimiter}{character(s) used as delimiter}

\item{quotes}{add quotes to each string}
}
\value{
collapsed string
}
\description{
Collapse vector of strings into single string with delimiters (and quotes)
}
