% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclicalMean.r
\name{cyclicalMean}
\alias{cyclicalMean}
\title{Cyclical mean}
\usage{
cyclicalMean(x, steps = 360)
}
\arguments{
\item{x}{vector of cyclical values}

\item{steps}{number of steps in the cycle}
}
\value{
collapsed string
}
\description{
Calculate the mean of a set of cyclical values, e.g. calendar months, day of year, angles. The code transform the cycle to cartesian vectors, calculates summed vector, and get the cycle of the sum vector.
}
