% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googledrive-to-local.r
\name{downDriveDir}
\alias{downDriveDir}
\title{Download Google Drive folder contents}
\usage{
downDriveDir(drivePath, localPath, authFile, rmFromDrive = FALSE)
}
\arguments{
\item{drivePath}{Path to directory on Google Drive that contains the files to download.}

\item{localPath}{Path where the files will be stored. Will be created if inexistent. No warnings if exists.}

\item{authFile}{Path to RDS file with authentication token. If not exist, web authentication is performd and new file is created at path.}

\item{rmFromDrive}{Remove the files from Google Drive after download (logical).}
}
\value{
Nothing
}
\description{
Download content of entire folder on Google Drive to a local directory
}
