% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{glaciers}
\alias{glaciers}
\title{Glacier data.}
\format{A data frame with 303 rows and 14 variables:
\describe{
  \item{\code{longitude}}{Cell centroid longitude}
  \item{\code{latitude}}{Cell centroid latitude}
  \item{\code{region}}{RGI region at cell centroid.}
  \item{\code{count}}{Number of glaciers in cell}
  \item{\code{precip}}{Mean annual precipitation from ERA-Interim for 1996-2015 (mm)}
  \item{\code{ela}}{Mean equilibrium line altitude (ELA)(m)}
  \item{\code{area_total}}{Total glacier area (km2)}
  \item{\code{area_ela}}{Glacier area below the ELA (km2)}
  \item{\code{area_debris}}{Debris covered glacier area (km2)}
  \item{\code{area_debris_ela}}{Debris covered glacier area below the ELA (km2)}
  \item{\code{mass_total}}{Total glacier mass (Gt)}
  \item{\code{mass_ela}}{Glacier mass below the ELA (Gt)}
  \item{\code{mass_debris}}{Debris covered glacier mass (Gt)}
  \item{\code{mass_debris_ela}}{Debris covered glacier mass below the ELA (Gt)}
}}
\source{
\href{https://www.nature.com/articles/nature23878}{Kraaijenbrink et al., 2017}
}
\usage{
glaciers
}
\description{
A dataset with glacier summaries for the high mountains of Asia on a 1x1 degree grid.
Data is limited to glaciers larger than 0.4 square km.
}
\keyword{datasets}
