% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{goldenratio}
\alias{goldenratio}
\title{Golden ratio.}
\format{An object of class \code{numeric} of length 1.}
\usage{
goldenratio
}
\description{
\href{https://en.wikipedia.org/wiki/Golden_ratio}{Wikipedia}
}
\keyword{datasets}
