% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-lines-to-inches.r
\name{linesToInches}
\alias{linesToInches}
\title{Convert lines to inches}
\usage{
linesToInches(x = 1)
}
\arguments{
\item{x}{Number of lines.}
}
\value{
Number of inches
}
\description{
Function to convert device lines to inches
}
