% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-color-transparant.r
\name{makeTransparent}
\alias{makeTransparent}
\title{Make color transparant}
\usage{
makeTransparent(colors = c("red", "green", "blue"), alpha = 0.5)
}
\arguments{
\item{alpha}{transparency value}

\item{x}{(vector of) color(s)}
}
\value{
transparant (vector of) color(s)
}
\description{
Make color transparant
}
