% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replot_axis_lines_no_ticks.r
\name{overplot.axes}
\alias{overplot.axes}
\title{Replot axis lines}
\usage{
overplot.axes(...)
}
\description{
Overplot the axis lines without any labels or ticks.
Useful when lines or polygons were plotted over the original axes.
}
