% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkRamp.r
\name{pal}
\alias{pal}
\alias{pkPal}
\title{Load preset color palettes and ramps}
\usage{
pal(name = "MntHydro", ...)

pkPal(name = "MntHydro", ...)
}
\arguments{
\item{name}{Name of the color ramp (string).}

\item{...}{Other arguments passed to \code{pkrf::ramp()}}
}
\value{
Vector with hex colors strings.
}
\description{
Load color ramps and palettes from presets. Identical to \code{pkrf::ramp()} besides default pallete.
}
