% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-colorbar-in-separate-plot.R
\name{pkColorbar}
\alias{pkColorbar}
\title{Plot color bar}
\usage{
pkColorbar(breaks = seq(0, 1, 0.1), colors = gray.colors(length(breaks)
  - 1), barlength = 0.5, barlwd = 0.5, axislab = "Colorscale",
  axislabline = 1.8, locut = F, hicut = F, smallticks = F,
  axisbreaks = breaks, axisbreaklab = axisbreaks, tck1 = -0.015,
  tck2 = -0.03, axiscex = 0.7, labelcex = 1, pmgp = c(2.5, 0.66,
  0), pmar = c(3, 0.5, 0.5, 0.5), ...)
}
\arguments{
\item{breaks}{vector of break values}

\item{colors}{vector of color}

\item{barlength}{relative length of the bar}

\item{barlwd}{line width for color bar}

\item{axislab}{axis label for the colorbar}

\item{axislabline}{spacing for the axis label}

\item{locut}{should the lower limit represent a cutoff? (boolean)}

\item{hicut}{should the higher limit represent a cutoff? (boolean)}

\item{smallticks}{plot small unlabeled ticks for each break (boolean)}

\item{axisbreaks}{labelled break values to use in the colorbar axis}

\item{axisbreaklab}{labels for the axisbreaks}

\item{tck1}{length of small ticks}

\item{axiscex}{cex for tick label}

\item{labelcex}{cex for axis label}

\item{pmgp}{temporary par 'mgp' to use}

\item{pmar}{temporary par 'mar' to use}

\item{...}{other arguments passed to axis function}

\item{tck1}{length of large ticks}
}
\value{
plotted colorbar
}
\description{
Plot color bar
}
