% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkLineToInch.r
\name{pkLineToInch}
\alias{pkLineToInch}
\title{Convert lines to inches}
\usage{
pkLineToInch(x = 1, cm = F)
}
\arguments{
\item{x}{Number of lines}

\item{cm}{Return centimeter instead (logical)}
}
\value{
Number of inches
}
\description{
Function to convert device lines to inches
}
