% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset-color-palette-generator.r
\name{pkPal}
\alias{pkPal}
\title{Load preset color palettes}
\usage{
pkPal(name = "Set1", number = 0, reversed = F, random = F)
}
\arguments{
\item{name}{Name of the palette (string).}

\item{number}{Number of output colors desired (integer).}

\item{reversed}{Reverse palette order (logical).}

\item{reversed}{Randomize palette order (logical).}
}
\value{
Vector with hex colors strings.
}
\description{
Load color palettes from preset
}
