% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset-color-ramp-generator.r
\name{pkRamp}
\alias{pkRamp}
\title{Load preset color ramps}
\usage{
pkRamp(name = "parula", number = 100, reversed = F)
}
\arguments{
\item{name}{Name of the color ramp (string).}

\item{number}{Number of output colors desired (integer).}

\item{reversed}{Should the ramp be reversed (logical).}
}
\value{
Vector with hex colors strings.
}
\description{
Load color ramps from preset
}
