% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkShadowText.r
\name{pkShadowText}
\alias{pkShadowText}
\title{Plot shadowed text labels}
\usage{
pkShadowText(x, y = NULL, labels, col = "white", bg = "black",
  theta = seq(pi/4, 2 * pi, length.out = 8), r = 0.1, ...)
}
\arguments{
\item{x}{x position}

\item{y}{y position}

\item{labels}{text labels}

\item{col}{foreground color}

\item{bg}{shadow/outline color}

\item{theta}{angle at which to plot the shadow}

\item{r}{distance of the shadow}

\item{...}{any other arguments for text function}
}
\value{
Vector with hex colors strings.
}
\description{
Plot text labels with a different color shadow or faked outline. Function taken from `TeachingDemos` package.
}
