% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkPlotB64.r
\name{plotB64}
\alias{plotB64}
\alias{pkB64}
\title{Plot byte64}
\usage{
plotB64(
  pfun,
  ...,
  width = 200,
  height = 200,
  res = 100,
  prefix = "<img src=\\"data:image/png;base64, ",
  suffix = "\\">"
)

pkB64(
  pfun,
  ...,
  width = 200,
  height = 200,
  res = 100,
  prefix = "<img src=\\"data:image/png;base64, ",
  suffix = "\\">"
)
}
\arguments{
\item{pfun}{a function that produces a plot}

\item{...}{other parameters passed to <code>pfun</code>}

\item{width}{PNG width (px)}

\item{height}{PNG width (px)}

\item{res}{PNG nominal resolution (ppi)}

\item{prefix}{prefix to apply to the byte64 string, useful to generate full HTML img src}

\item{suffix}{suffix to apply to the byte64 string, useful to generate full HTML img src}
}
\value{
string with byte64 data
}
\description{
Plot image to temporary png and load as Byte64 data
}
