% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkRastplot.r
\name{plotRast}
\alias{plotRast}
\alias{pkRastplot}
\title{Plot raster with limits}
\usage{
plotRast(x, lcut = NA, hcut = NA, setNA = NA, useimage = F, ...)

pkRastplot(x, lcut = NA, hcut = NA, setNA = NA, useimage = F, ...)
}
\arguments{
\item{x}{raster}

\item{lcut}{Minimum value to plot, all below will be similarly coloured. Set to \code{NA} to disable.}

\item{hcut}{Maximum value to plot, all above will be similarly coloured. Set to \code{NA} to disable.}

\item{setNA}{Value in raster to set as \code{NA} in output. Set to \code{NA} to disable.}

\item{useimage}{Use \code{image.plot} to make the plot rather than \code{plot.raster} (logical).}
}
\value{
no function return
}
\description{
Base plot of raster, with inlusion of lower and higher plot limits
}
