% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkShadeplot.r
\name{plotRibbons}
\alias{plotRibbons}
\alias{pkShadeplot}
\title{Plot shaded polygons}
\usage{
plotRibbons(x, y, clr, ...)

pkShadeplot(x, y, clr, ...)
}
\arguments{
\item{x}{Data frame with x values}

\item{y}{Data frame with y values}

\item{clr}{vector with colors}

\item{...}{Other parameters parsed to polygon function}
}
\value{
Shaded areas in current plot
}
\description{
Plot shaded polygons between any number of input lines
}
