% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-raster-over-rdp.r
\name{rrplot}
\alias{rrplot}
\title{Raster plot over RDP}
\usage{
rrplot(x, ...)
}
\arguments{
\item{x}{Raster as defined by the \code{"raster"} package.}

\item{...}{Other parameters parsed to the raster plot function.}
}
\value{
Plot of the raster that is correctly displayed
}
\description{
Function to plot rasters correctly over an Windows remote desktop protocol connection.
}
