% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scientific_clean.r
\name{scientific10}
\alias{scientific10}
\title{Clean scientific notation labeler}
\usage{
scientific10(x, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{...}{Any other argument passed to \code{scales::scientific()}}
}
\value{
Expression vector
}
\description{
Convert numbers into a scientific notation expression that uses \out{<i>2.4 x 10<sup>-3</sup></i>} format
}
