% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheet.r
\name{sheet}
\alias{sheet}
\title{Cheat sheets for various topics}
\usage{
sheet(topic = "topics")
}
\arguments{
\item{topic}{Topic to display cheat sheet for. Defaults to 'topics', which will provide a list of available topics.}
}
\description{
Display a pdf cheat sheet for various R topics and packages
}
\examples{
sheet('topics')
sheet('ggplot2')

}
