% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mh_minimal.r
\name{theme_mh_minimal}
\alias{theme_mh_minimal}
\title{Custom ggplot2 theme PK style}
\usage{
theme_mh_minimal(base_size = 10, base_family = "",
  base_line_size = base_size/22, base_rect_size = base_size/22,
  alwd = 0.75, grid = F, yrot = F, strip = "dark")
}
\arguments{
\item{base_size}{base point size}

\item{base_family}{base font family (best to use \code{showtext} package and Segoe UI fonts)}

\item{base_line_size}{= base line size}

\item{base_rect_size}{= base rect size}

\item{grid}{draw no (\code{0}), only major ((\code{1})) or major and minor grid lines (\code{2})}

\item{yrot}{rotate y axis tick labelsl 90 degrees (boolean)}

\item{strip}{Use dark blue facet strip with white text ("dark") or white strip with blue text ("light")}
}
\value{
\code{ggplot2} theme opbject
}
\description{
Custom ggplot2 theme PK style
}
