% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pk.r
\name{theme_pk}
\alias{theme_pk}
\title{Custom ggplot2 theme}
\usage{
theme_pk(base_size = 10, base_family = "",
  base_line_size = base_size/22, base_rect_size = base_size/22,
  alwd = 0.5, grid = F, yrot = F)
}
\arguments{
\item{base_size}{base point size}

\item{base_family}{base font family}

\item{base_line_size}{= base line size}

\item{base_rect_size}{= base rect size}

\item{grid}{draw major and minor grid lines (boolean)}

\item{yrot}{rotate y axis tick labelsl 90 degrees (boolean)}
}
\description{
Custom ggplot2 theme
}
