% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkTintShade.r
\name{tint}
\alias{tint}
\alias{pkTintShade}
\title{Shade or tint}
\usage{
tint(colors = c("red", "green", "blue"), opt = "shade", factor = 0.5)

pkTintShade(colors = c("red", "green", "blue"), opt = "shade", factor = 0.5)
}
\arguments{
\item{opt}{"tint" (lighten) or "shade" (darken)}

\item{x}{(vector of) color(s) with no alpha}
}
\value{
altered (vector of) color(s)
}
\description{
Shade or tint
}
