% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-color-shade-or-tint.r
\name{tintShade}
\alias{tintShade}
\title{Shade or tint}
\usage{
tintShade(colors = c("red", "green", "blue"), opt = "shade",
  factor = 0.5)
}
\arguments{
\item{opt}{"tint" (lighten) or "shade" (darken)}

\item{x}{(vector of) color(s) with no alpha}
}
\value{
altered (vector of) color(s)
}
\description{
Shade or tint
}
