% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unixPath.r
\name{unixPath}
\alias{unixPath}
\title{Convert Windows to Unix paths}
\usage{
unixPath(path, reverse = FALSE, mountpoint = "/mnt")
}
\arguments{
\item{path}{Path to be transformed.}

\item{reverse}{Convert from windows to unix (\code{FALSE}) or from unix to windows (\code{TRUE}). When converting from unix, a full path must be provided. Converting to unix can be performed with relative paths.}

\item{mountpoint}{Mountpoint of the windows drive, usually \code{'/mnt'} for linux subsystem and \code{'/cygdrive'} for cygwin.}
}
\value{
transformed path
}
\description{
Get unix paths equivalents of Windows paths, for example for linux subsystem or cygwin. Function likely only applicable under Windows.
}
