% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkHeatmap.r
\name{heatGrid}
\alias{heatGrid}
\alias{pkHeatmap}
\title{Create heat map}
\usage{
heatGrid(mat = NULL, xrange = NULL, yrange = NULL, xlimit = NULL,
  ylimit = NULL, colramp = topo.colors(200), maketicks = T,
  makebox = T, axislwd = 0.5, locut = NULL, hicut = NULL,
  nacut = F, ...)

pkHeatmap(mat = NULL, xrange = NULL, yrange = NULL, xlimit = NULL,
  ylimit = NULL, colramp = topo.colors(200), maketicks = T,
  makebox = T, axislwd = 0.5, locut = NULL, hicut = NULL,
  nacut = F, ...)
}
\arguments{
\item{mat}{matrix with values to create heatmap for}

\item{xrange}{the x range in which to plot the grid (vector length 2)}

\item{yrange}{the y range in which to plot the grid (vector length 2)}

\item{xlimit}{the x limits of the plot (vector length 2)}

\item{ylimit}{the y limits of the plot (vector length 2)}

\item{colramp}{the colors to use for plotting (vector)}

\item{maketicks}{draw axis ticks and tick marks (logical)}

\item{makebox}{draw box line around plot (logical)}

\item{axislwd}{linewidth to use for ticks and box (numeric)}

\item{locut}{low threshold on raster values (numeric)}

\item{hicut}{high threshold on raster values (numeric)}

\item{nacut}{values outside cutrange are set to NA instead of cut value (logical)}

\item{...}{other parameters passed to image/plot/axis functions}
}
\value{
plot
}
\description{
Function for a simple heat map with custom limits in base r plotting. The function is a wrapper for the image function
}
